+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Wed, 26 Jul 2023 21:12:23 GMT and should not be manually modified.
+This log was last generated on Tue, 01 Aug 2023 16:43:05 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 17.0.3
+
+Tue, 01 Aug 2023 16:43:05 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v20.0.3
+
+## 17.0.2
+
+Mon, 31 Jul 2023 22:52:43 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v20.0.2
 
 ## 17.0.1